Make chaincode average calculator chan buffered
This modification was made since we detected that non-buffered
go channels block until the receiver reads, which is what we DID
NOT wanted. To correct this, we made the channel buffer size
be equal to the CORE_PEER_LIMITS_CONCURRENCY_ENDORSERSERVICE
configuration.

We made the peer create the chaincode with that environment
variable set, so the chaincode can read it and create the
buffered channel with the respective size.